Added func dynamicType that represents the type() aux-function (paper)
Performed some minor corrections related to the introduction of dynamicType.

Some corrections are also related to the fact that some code used to assume that there were only 2 types: u_S and u_I. This assumption led to writing isStructType to mean !isInterfaceType, but after introducing primitive types and general type defs that is no longer the case.