Refactored TSubs,EtaOpen into EtaOpen; Eta into EtaClosed
Idea: both TSubs and SubsEtaOpen are applied in cases where there may be no mapping for a given type variable (or the mapping itself contaings type vars). Hence merged TSubs and SubsEtaOpe into a single SubsEtaOpen.
On the other hand, SubsEta is applied when every variable is 'mappable' to some ground type (given Delta), thus was renamed to SubsEtaClosed.