[WIP] AssignableTo with "coercion semantics"
Essentially, AssignableTo having coercion semantics means that this method now returns both a boolean and a Coercion function. A coercion function is of type FGExpr -> FGExpr (for fg). 
The idea is that, when an AssignableTo test of the kind <literal>.AssignableTo<defined type> succeeds, we also get a function that knows how to turn a Literal-ASTNode into a node of the expected type.

I believe these coercions are a good fit to model Go's behavior. Moreover, some form of coercion is actually necessary to satisfy type-safety at each small step of evaluation. For some evidences of this claim, check the committed examples at fg_prims_test.go.